+++ src/main/java/com/dashboard/repository/AnggotaKliringRepo.java
@@ -7,7 +7,7 @@ import com.dashboard.model.AnggotaKliring;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AnggotaKliringRepo extends JpaRepository<AnggotaKliring, Long>{
+public interface AnggotaKliringRepo extends JpaRepository<AnggotaKliring, Integer>{
     @Query("FROM AnggotaKliring WHERE id = ?1")
-    List<AnggotaKliring> findAnggotaKliringById(Long id);
+    List<AnggotaKliring> findAnggotaKliringById(Integer id);
 }
